Moved code into its own component
