+++ components/Page.js
@@ -3,32 +3,12 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import styled, { ThemeProvider } from 'styled-components';
-import Router from 'next/router';
 import Header from './Header';
 import Footer from './Footer';
 import Meta from './Meta';
 import GlobalStyle from '../styles/GlobalStyle';
 import theme from '../theme';
-import { initGA, LogPageView } from './GoogleAnalytics';
-
-// Fetch env variables
-async function fetchData(url) {
-  const res = await fetch(url);
-  const params = await res.json();
-  return params;
-}
-
-Router.events.on('routeChangeStart', url => {
-  console.log(`Loading: ${url}`);
-
-  fetchData(`${window.location.origin}/.netlify/functions/fetch_env_vars`)
-    .then(params => {
-      initGA(params.analyticsID);
-    })
-    .then(() => {
-      LogPageView();
-    });
-});
+import GoogleAnalytics from './GoogleAnalytics';
 
 class Page extends Component {
   constructor(props) {
@@ -36,17 +16,6 @@ class Page extends Component {
 
     this.state = { headerMenuOpen: false };
     this.onHeaderMenuToggle = this.onHeaderMenuToggle.bind(this);
-  }
-
-  componentDidMount() {
-    const url = `${window.location.origin}/.netlify/functions/fetch_env_vars`;
-    fetchData(url)
-      .then(params => {
-        initGA(params.analyticsID);
-      })
-      .then(() => {
-        LogPageView();
-      });
   }
 
   onHeaderMenuToggle(menuHidden) {
@@ -58,6 +27,7 @@ class Page extends Component {
       <ThemeProvider theme={theme}>
         <StyledPage className={this.state.headerMenuOpen ? 'header-menu-open' : ''}>
           <Meta />
+          <GoogleAnalytics />
           <GlobalStyle />
           <Header onMenuToggle={this.onHeaderMenuToggle} theme={theme} />
           <div>{this.props.children}</div>
